+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsGroupPolicyAssignment/MSFT_TeamsGroupPolicyAssignment.psm1
@@ -121,7 +121,7 @@ function Get-TargetResource
         Write-Verbose -Message $Message
         return @{
             GroupId               = $Group.Id
-            GroupDisplayName      = $Group.Displayname
+            GroupDisplayName      = $Group.DisplayName
             PolicyType            = $GroupPolicyAssignment.PolicyType
             PolicyName            = $GroupPolicyAssignment.PolicyName
             Priority              = $GroupPolicyAssignment.Priority
@@ -443,7 +443,7 @@ function Export-TargetResource
                 ManagedIdentity       = $ManagedIdentity.IsPresent
                 AccessTokens          = $AccessTokens
             }
-            #$results = Get-TargetResource @getParams
+            #$results = Get-TargetResource @results
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `